refactor: merge affinidi-sd-jwt-vc into affinidi-vc as sd_jwt_vc module (W18b)
SD-JWT VC is a credential format, so its implementation now lives alongside the
W3C VC data model in affinidi-vc rather than in its own crate.

- affinidi-vc 0.1.2 → 0.2.0: new `sd_jwt_vc` module (moved verbatim from the
  old crate, history preserved via git mv for error.rs), depends on
  affinidi-sd-jwt, adds a `_test-utils` passthrough feature; a dev-dependency on
  affinidi-sd-jwt/_test-utils keeps the module's tests compiling under a plain
  `cargo test`. All 36 vc tests pass (incl. the 11 moved sd_jwt_vc tests).
- affinidi-sd-jwt-vc 0.1.1 → 0.2.0: now a deprecated thin re-export of
  affinidi_vc::sd_jwt_vc (incl. the `error` submodule), kept for one release.
  Every former `affinidi_sd_jwt_vc::*` path still resolves. Dropped its now-unused
  deps (chrono/serde/serde_json/thiserror/tracing); only depends on affinidi-vc.
- affinidi-tdk 0.8.1 → 0.8.2: `sd-jwt-vc` feature now implies `vc`, and
  affinidi_tdk::sd_jwt_vc re-exports affinidi_vc::sd_jwt_vc — the facade path is
  unchanged for users and the dep on the old crate is dropped.

No behaviour, signatures, or error variants changed — only the import path.
None of these crates is externally [patch.crates-io]-pinned, so normal semver
bumps apply (ADR-0003 patch-not-minor constraint does not bite here).

Migration: docs/migration/2026-06-sd-jwt-vc-merge.md; CHANGELOG [Unreleased].
Verified: fmt clean; clippy -D warnings green on vc/shim/facade; facade builds
under --features sd-jwt-vc and --features credentials.
