+++ crates/credentials/affinidi-sd-jwt-vc/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "affinidi-sd-jwt-vc"
-description = "SD-JWT-based Verifiable Credentials (SD-JWT VC) per IETF draft-ietf-oauth-sd-jwt-vc."
-version = "0.1.1"
+description = "DEPRECATED — merged into affinidi-vc; use affinidi_vc::sd_jwt_vc. Thin re-export kept for one release."
+version = "0.2.0"
 edition.workspace = true
 authors.workspace = true
 readme = "README.md"
@@ -12,19 +12,7 @@ repository.workspace = true
 publish.workspace = true
 rust-version.workspace = true
 
-[features]
-default = []
-_test-utils = ["affinidi-sd-jwt/_test-utils"]
-
+# DEPRECATED: this crate re-exports affinidi_vc::sd_jwt_vc. All implementation
+# (and its tests) moved to affinidi-vc; the only dependency is the new home.
 [dependencies]
-affinidi-sd-jwt = { version = "0.1", path = "../affinidi-sd-jwt" }
-
-chrono = { version = "0.4", features = ["serde"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-thiserror = "2"
-tracing = "0.1"
-
-[dev-dependencies]
-affinidi-sd-jwt-vc = { path = ".", features = ["_test-utils"] }
-tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+affinidi-vc = { version = "0.2", path = "../affinidi-vc" }
