+++ crates/credentials/affinidi-vc/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "affinidi-vc"
 description = "W3C Verifiable Credentials Data Model 1.1 and 2.0 implementation."
-version = "0.1.2"
+version = "0.2.0"
 edition.workspace = true
 authors.workspace = true
 readme = "README.md"
@@ -12,7 +12,17 @@ repository.workspace = true
 publish.workspace = true
 rust-version.workspace = true
 
+[features]
+default = []
+# Re-export of affinidi-sd-jwt's test signer/hasher helpers, used by the
+# `sd_jwt_vc` module's tests and available to downstream test code.
+_test-utils = ["affinidi-sd-jwt/_test-utils"]
+
 [dependencies]
+# Backs the `sd_jwt_vc` module (merged in from the former affinidi-sd-jwt-vc
+# crate). SD-JWT VC is a credential format, so it lives with the VC data model.
+affinidi-sd-jwt = { version = "0.1", path = "../affinidi-sd-jwt" }
+
 chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -21,4 +31,7 @@ tracing = "0.1"
 url = "2"
 
 [dev-dependencies]
+# Enable the SD-JWT test signer/hasher for this crate's own test builds so the
+# `sd_jwt_vc` tests compile under a plain `cargo test` (no feature flag needed).
+affinidi-sd-jwt = { version = "0.1", path = "../affinidi-sd-jwt", features = ["_test-utils"] }
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
