+++ crates/credentials/affinidi-vc/src/sd_jwt_vc/error.rs
